+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Add `rodauth:migration` generator for creating tables of specified features (@janko)
+
 * Use UUIDs for primary keys if so configured in Rails generators (@janko)
 
 * Add `rodauth:routes` rake task for printing routes handled by Rodauth middleware (@janko)